requirements.txt
flask>=3.1
pyjwt>=2.10
requests>=2.31
cryptography>=46.0
